pull out _calculate_points method into shared location
